Clamp the boundary phase bucket on read, as the writer already did
The phase bucketing of the boundary union is computed twice with the same
formula: BoundaryUnion1D::bucketIndex sizes and fills the bucket row,
BoundaryViolationDetector::phaseBucket indexes it back. The writer clamps
to its last bucket; the reader did not.

The Nichols window is free text in the boundaries dialog, validated only
for ordering, while every caller normalises a loop phase into (-360, 0].
A window narrower than 360 degrees - [-180, 0] is a reasonable zoom - is
enough for a phase below its start to index past the end of the row.
pointVerdict reads it with at(), which is undefined behaviour out of
range, and the box classification reaches the same row with value(),
which answers nullptr and was then dereferenced. A test reproduces the
segmentation fault.

Note for the review: the clamp stops the crash but gives a point outside
the window the verdict of the edge bucket, which nobody computed. For a
search that claims a global optimum that can be worse than an error, so
whether this should reject the window instead - and where, the dialog or
the search - is left for the joint review.

Also, from a pass with an aggressive warning set over the 53 translation
units: 10 signed/unsigned comparisons and one stray semicolon, all of
them emitted by the -Wall the project already builds with. The build was
not warning free; the incremental builds I had been checking simply
compiled nothing. Two parameters carrying the m_ prefix of the member
they feed are renamed, that being the kind of shadowing that hid the
stylesheet defect.

185 tests pass.

+++ GUI/loop_shaping_viewer.cpp
@@ -78,10 +78,10 @@ void LoopShapingViewer::showDiagram(){
     QString numerador = "", denominador = "";
 
     qint32 i = 0;
-    for (i = 0; i < loopShapingData->controller()->numerator().size(); i++){
+    for (i = 0; i < static_cast<qint32>(loopShapingData->controller()->numerator().size()); i++){
         numerador += QString::number(loopShapingData->controller()->numerator()[i].nominal()) + " ";
     }
-    for (i = 0; i < loopShapingData->controller()->denominator().size(); i++){
+    for (i = 0; i < static_cast<qint32>(loopShapingData->controller()->denominator().size()); i++){
         denominador += QString::number(loopShapingData->controller()->denominator()[i].nominal()) + " ";
     }
